requirements.txt
numpy==2.2.0
pandas==2.2.3
py_mini_racer==0.6.0
mini-racer==0.12.4
TA_Lib==0.5.1
arrow==1.3.0
bokeh==3.6.2
PyMySQL==1.1.1
requests==2.32.3
Logbook==1.8.0
SQLAlchemy==2.0.36
tornado==6.4.2
tqdm==4.67.1
easytrader==0.23.0
beautifulsoup4==4.12.3
pycryptodome==3.21.0
python_dateutil==2.9.0.post0
backtrader~=1.9.78.123
matplotlib~=3.9.3
python-dateutil~=2.9.0.post0
easyquotation~=0.7.5